+++ album.html
@@ -11,7 +11,6 @@
     <div id="album-item-wrapper"></div>
     <script type="module" src="album.js"></script>
     <script type="module" src="./export.js"></script>
-    <script type="module" src="header.js"></script>
-    <script src="functions.js"></script>
+   
 </body>
 </html>
