+++ albums.html
@@ -11,9 +11,7 @@
     <div id="albums-wrapper"></div>
     
     <script type="module" src="./albums.js"></script>
-    <script type="module" src="./header.js"></script>
    
-    <script src="functions.js"></script>
     
     <!-- <script src="./functions.js"></script> -->
 </body>
