+++ index.html
@@ -14,7 +14,8 @@
     <div id="album-wrapper"></div>
     
     <script type="module" src="script.js"></script>
-    <script type="module" src="header.js"></script>
+   
     <script type="module" src="functions.js"></script>
+    <script type="module" src="./export.js"></script>
 </body>
 </html>
