+++ post.html
@@ -9,7 +9,7 @@
 </head>
 <body>
     <div id="post-wrapper"></div>
-    <script src="post.js"></script>
-    <script src="header.js"></script>
+    <script type="module" src="post.js"></script>
+   
 </body>
 </html>
