+++ post.js
@@ -1,3 +1,5 @@
+import headerElement from "./header.js"
+headerElement()
 let queryParams = document.location.search;
 let urlParams = new URLSearchParams(queryParams);
 let postId = urlParams.get('post_id');
