+++ posts.html
@@ -10,8 +10,8 @@
 <body>
       <h1>Posts page:</h1>
       <div id="post-wrapper"></div>
-    <script type="module" src="functions.js"></script>
+ 
     <script type="module" src="posts.js"></script>
-    <script type="module" src="header.js"></script>
+   
 </body>
 </html>
