+++ script.js
@@ -1,5 +1,6 @@
 import{renderListElement,toUpperCase} from "./export.js"
-
+import  headerElement from "./header.js"
+headerElement()
 const postWrapper = document.getElementById('post-wrapper')
 
     function init() {
