+++ search.html
@@ -18,6 +18,6 @@
     
     <script src="functions.js"></script>
     <script src="header.js"></script>
-    <script src="search.js"></script>
+    <script type="module" src="search.js"></script>
 </body>
 </html>
