+++ user.html
@@ -12,8 +12,8 @@
     <div id="post-wrapper"></div>
     <div id="album-wrapper"></div>
 
-    <script   src="user.js"></script>
-    <script   src="header.js"></script>
+   
+    <script type="module"   src="user.js"></script>
     <script  src="functions.js"></script>
     
 </body>
